+++ conjunction/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Upcoming <!-- Add unreleased change notes here: -->
 
+- Remove unused Spicerack dependency ([#460](https://github.com/Freshly/spicerack/pull/460))
+
 ## v0.26.0.3
 
 *Release Date*: 5/4/2021